+++ CHANGELOG.md
@@ -62,6 +62,7 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 * Include `railway` and `building` features with lifecycle status only in _past/future_ features category ([#9058])
 * Disable autocomplete on changeset comment box on mobile Firefox to work around buggy/glitchy behavior of the input field ([#9011])
 * Fix crash when trying to delete a single-node way ([#9007])
+* Make sure the available options of multi/semi-combo fields are updated properly when a previous tag change caused a change of the feature's preset ([#12217])
 #### :earth_asia: Localization
 #### :hourglass: Performance
 #### :mortar_board: Walkthrough / Help
@@ -96,6 +97,7 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 [#12139]: https://github.com/openstreetmap/iD/pull/12139
 [#12169]: https://github.com/openstreetmap/iD/issues/12169
 [#12178]: https://github.com/openstreetmap/iD/pull/12178
+[#12217]: https://github.com/openstreetmap/iD/issues/12217
 
 
 # 2.39.6